Move back methods which are not needed in their new locations
Moved back:
- gameManager.getAllMarkableQuestionPartsDFSOrder
- gameManager.filterQuestionParts
- gameManager.depthFirstQuestionSearch

Kept:
- statisticsManager.getQuestionMap -> questionAttemptsManager.getQuestionMap. Needed to avoid a statisticsManager <-> gameManager dependency loop. Additionally getQuestionMap doesn't really belong in statisticsManager.
