+++ src/main/java/uk/ac/cam/cl/dtg/segue/api/managers/QuestionManager.java
@@ -29,7 +29,6 @@ import uk.ac.cam.cl.dtg.isaac.configuration.IsaacApplicationRegister;
 import uk.ac.cam.cl.dtg.isaac.dos.TestCase;
 import uk.ac.cam.cl.dtg.isaac.dos.TestQuestion;
 import uk.ac.cam.cl.dtg.isaac.dto.IsaacItemQuestionDTO;
-import uk.ac.cam.cl.dtg.isaac.dto.IsaacQuickQuestionDTO;
 import uk.ac.cam.cl.dtg.segue.api.Constants;
 import uk.ac.cam.cl.dtg.segue.api.Constants.*;
 import uk.ac.cam.cl.dtg.segue.dao.SegueDatabaseException;
@@ -644,72 +643,5 @@ public class QuestionManager {
         }
 
         return questionIdToQuestionMap;
-    }
-
-    /**
-     * Get all questions in the question page: depends on each question. This method will conduct a DFS traversal and
-     * ensure the collection is ordered as per the DFS.
-     *
-     * @param questionPageId
-     *            - results depend on each question having an id prefixed with the question page id.
-     * @return collection of markable question parts (questions).
-     * @throws ContentManagerException
-     *             if there is a problem with the content requested.
-     */
-    public Collection<QuestionDTO> getAllMarkableQuestionPartsDFSOrder(final String questionPageId)
-            throws ContentManagerException {
-        Validate.notBlank(questionPageId);
-
-        // do a depth first traversal of the question page to get the correct order of questions
-        ContentDTO questionPage = this.contentManager.getContentById(this.contentManager.getCurrentContentSHA(),
-                questionPageId);
-        List<ContentDTO> dfs = Lists.newArrayList();
-        dfs = depthFirstQuestionSearch(questionPage, dfs);
-
-        return this.filterQuestionParts(dfs);
-    }
-
-    /**
-     * We want to list the questions in the order they are seen.
-     * @param c - content to search
-     * @param result - the list of questions
-     * @return a list of questions ordered by DFS.
-     */
-    private List<ContentDTO> depthFirstQuestionSearch(final ContentDTO c, final List<ContentDTO> result) {
-        if (c == null || c.getChildren() == null || c.getChildren().size() == 0) {
-            return result;
-        }
-
-        for (ContentBaseDTO child : c.getChildren()) {
-            if (child instanceof QuestionDTO) {
-                result.add((QuestionDTO) child);
-                // assume that we can't have nested questions
-            } else {
-                depthFirstQuestionSearch((ContentDTO) child, result);
-            }
-        }
-        return result;
-    }
-
-    /**
-     * Utility method to extract a list of questionDTOs only.
-     *
-     * @param contentToFilter
-     *            list of content.
-     * @return list of question dtos.
-     */
-    private Collection<QuestionDTO> filterQuestionParts(final Collection<ContentDTO> contentToFilter) {
-        List<QuestionDTO> results = Lists.newArrayList();
-        for (ContentDTO possibleQuestion : contentToFilter) {
-
-            if (!(possibleQuestion instanceof QuestionDTO) || possibleQuestion instanceof IsaacQuickQuestionDTO) {
-                // we are not interested if this is not a question or if it is a quick question.
-                continue;
-            }
-            QuestionDTO question = (QuestionDTO) possibleQuestion;
-            results.add(question);
-        }
-
-        return results;
     }
 }
